Add tests and enhance functionality for track manipulation and file addition
This commit adds new unit tests for actions related to track manipulation and file addition in the MKVFile class. It also expands the functionality of add_file method to accept os.PathLike objects and enhances the cleanup process for test files. Changes in the Makefile are made for downloading additional test files.