Replace curl with wget in workflow setup
The setup process in the GitHub Actions workflow has been updated. The command for downloading GPG keys for the MKVToolNix repository was using curl, but it has been replaced with wget. This change can add more reliability as wget is more universal and better installed by default on some systems.