Adding linting checks with mypy.
Integrating with pytest.

Updating types.

updating siginature of order_tracks_by_file_id to raise ValueError if KeyError is raised to ensure the contract of track.file_id when a None value is assigned is preserved.

updating tests.yml to specify ubuntu version as the ubuntu-latest
[label](https://github.com/actions/runner-images/pull/10807)
changing affects which version of the mkvtoolnix
[repo](https://mkvtoolnix.download/downloads.html#ubuntu) to download from.